fix(types): defer typing to PyCharm when we have no override (#11)
This made no difference for most arguments, because we were returning
the same type that PyCharm would. However, we were improperly handling
*args: instead of returning `tuple[type]`, we were returning `type`.
This caused `args[0]` to be improper, in case of scalars, as well as
`for thing in args:` to beef it.

This commit simply defers to PyCharm's typing system when we have
nothing to add to it. Our previous behaviour was a change made to
_avoid_ overriding explicit types added to declared fixture names.
However, the proper move would have been to defer to PyCharm by
returning null, instead of telling PyCharm we know better by returning
an actual type reference.