fix(config): retrieve CommandLine lang dynamically, to avoid dupe reg issues from static instance
Mayhaps something changed with the class loaders to introduce more isolation between plugins. When referencing CommandLineLanguage.INSTANCE, an error would be thrown: Language with ID 'CommandLine' is already registered. It seems _my_ copy of CommandLineLanguage was not the same instance used to register the language initially (i.e. the two copies do not share the same static namespace), so when I accessed mine, it thought it hadn't registered, yet.